feat(210-04): narrow day-index uint48 to uint32 in 5 test files + handler
- LockRemoval.t.sol: rngWordByDay mapping keys uint48 -> uint32
- RedemptionGas.t.sol: currentDay variable uint48 -> uint32
- VRFPathCoverage.t.sol: loop vars and casts uint48 -> uint32 at 4 sites
- VRFStallEdgeCases.t.sol: loop vars uint48 -> uint32 at 5 sites
- VRFPathHandler.sol: dayAfter, gapStart, gapEnd, loop var uint48 -> uint32

+++ test/fuzz/LockRemoval.t.sol
@@ -30,7 +30,7 @@ contract LockRemovalHarness is DegenerusGameStorage {
     }
 
     function setRngWordByDay(uint256 day, uint256 word) external {
-        rngWordByDay[uint48(day)] = word;
+        rngWordByDay[uint32(day)] = word;
     }
 
     // --- LOCK-01: _callTicketPurchase guard (MintModule:838-840 post-removal) ---
@@ -63,7 +63,7 @@ contract LockRemovalHarness is DegenerusGameStorage {
 
     // --- LOCK-06: requestLootboxRng guard (AdvanceModule:641-644 post-removal, line 643 deleted) ---
     function requestLootboxRngGuard(uint256 currentDay) external view {
-        if (rngWordByDay[uint48(currentDay)] == 0) revert E();
+        if (rngWordByDay[uint32(currentDay)] == 0) revert E();
         // rngLockedFlag check REMOVED (was line 643)
         if (rngRequestTime != 0) revert E();
     }
