+++ test/fuzz/VRFPathCoverage.t.sol
@@ -158,7 +158,7 @@ contract VRFPathCoverage is DeployProtocol {
         uint256 gapCount = resumeDay - stallDay;
         uint256[] memory words = new uint256[](gapCount);
         for (uint256 d = stallDay; d < resumeDay; d++) {
-            uint256 w = game.rngWordForDay(uint48(d));
+            uint256 w = game.rngWordForDay(uint32(d));
             assertTrue(w != 0, "Gap day word must be nonzero");
             words[d - stallDay] = w;
         }
@@ -201,7 +201,7 @@ contract VRFPathCoverage is DeployProtocol {
         assertTrue(gasUsed < 25_000_000, "120-day gap backfill must use < 25M gas");
 
         // Verify all 120 gap days (3..122) have nonzero words
-        for (uint48 d = 3; d <= 122; d++) {
+        for (uint32 d = 3; d <= 122; d++) {
             assertTrue(
                 game.rngWordForDay(d) != 0,
                 "120-day gap: all gap days must have nonzero words"
@@ -253,7 +253,7 @@ contract VRFPathCoverage is DeployProtocol {
         _resumeAfterSwap(newVRF, vrfWord);
 
         // Verify gap days have nonzero words (days 4..7 are the gap)
-        for (uint48 d = 4; d <= 7; d++) {
+        for (uint32 d = 4; d <= 7; d++) {
             assertTrue(
                 game.rngWordForDay(d) != 0,
                 "Gap day must have nonzero word after mid-day stall recovery"
@@ -294,7 +294,7 @@ contract VRFPathCoverage is DeployProtocol {
 
         // Collect all gap day words (days 3..12)
         uint256[10] memory words;
-        for (uint48 d = 3; d <= 12; d++) {
+        for (uint32 d = 3; d <= 12; d++) {
             uint256 w = game.rngWordForDay(d);
             assertTrue(w != 0, "Gap day word must be nonzero");
             words[d - 3] = w;
